refactor: Move retry_on_exception decorator to subset selection utils
- Relocate retry_on_exception decorator from subset_selection.py to subset_selection_utils.py
- Remove unnecessary imports in subset_selection.py
- Simplify code structure and improve module organization
- Maintain existing error handling and retry logic
